Swirl attack
Implementation of the swirl/spinning attack of a throwing blade around Beff.

+++ SaveSystems/GameData.cs
@@ -9,7 +9,7 @@ public class GameData
 {
     public string name;
     public DateTime date;
-    public int timePlayed;
+    public double timePlayed;
 
     public int totalPoints;
     public int totalMoney;
